Rename types directory to delighters
